equestrian-fei-system/backend/requirements/base.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.1
redis==4.6.0
djangorestframework-simplejwt==5.3.0
django-filter==23.2
pillow==10.3.0
django-redis==5.3.0
reportlab==4.0.4
pandas==2.1.1
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.12.2